updated recipe list fragment

+++ app/src/main/java/com/example/cookbookapp/entity/Recipe.kt
@@ -10,5 +10,5 @@ data class Recipe(
     var ingredients: List<String>? = null,
     var info: String? = null,
     var id: String? = null,
-    var score: Double? = null
+    var score: Float? = null
 )
